move Conv2DBNBlock and add explicit padding option for the use of object detection use case
